+++ projects/igniteui-angular/src/lib/select/select.component.ts
@@ -341,6 +341,12 @@ export class IgxSelectComponent extends IgxDropDownComponent implements IgxSelec
         super.open(Object.assign({}, this._overlayDefaults, this.overlaySettings, overlaySettings));
     }
 
+    public inputGroupClick(event: MouseEvent) {
+        if ((event.target as HTMLElement).className !== 'igx-input-group__hint') {
+            this.toggle();
+        }
+    }
+
     /** @hidden @internal */
     ngAfterContentInit() {
         this._overlayDefaults = {